+++ dom/elements/src/main/java/DOMGetAttributes.java
@@ -21,20 +21,19 @@
 import com.teamdev.jxbrowser.chromium.Browser;
 import com.teamdev.jxbrowser.chromium.dom.By;
 import com.teamdev.jxbrowser.chromium.dom.DOMDocument;
-import com.teamdev.jxbrowser.chromium.dom.DOMOptionElement;
-import com.teamdev.jxbrowser.chromium.dom.DOMSelectElement;
+import com.teamdev.jxbrowser.chromium.dom.DOMElement;
 import com.teamdev.jxbrowser.chromium.events.FinishLoadingEvent;
 import com.teamdev.jxbrowser.chromium.events.LoadAdapter;
 import com.teamdev.jxbrowser.chromium.swing.BrowserView;
 
 import javax.swing.*;
 import java.awt.*;
-import java.util.List;
+import java.util.Map;
 
 /**
- * This example demonstrates how to programatically select an option item in SELECT tag.
+ * Demonstrates how to get list of existing attributes of a specified HTML element.
  */
-public class DOMSelectOptionExample {
+public class DOMGetAttributes {
     public static void main(String[] args) {
         Browser browser = new Browser();
         BrowserView browserView = new BrowserView(browser);
@@ -42,7 +41,7 @@ public class DOMSelectOptionExample {
         JFrame frame = new JFrame();
         frame.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
         frame.add(browserView, BorderLayout.CENTER);
-        frame.setSize(700, 500);
+        frame.setSize(800, 600);
         frame.setLocationRelativeTo(null);
         frame.setVisible(true);
 
@@ -50,23 +49,15 @@ public class DOMSelectOptionExample {
             @Override
             public void onFinishLoadingFrame(FinishLoadingEvent event) {
                 if (event.isMainFrame()) {
-                    Browser browser = event.getBrowser();
-                    DOMDocument document = browser.getDocument();
-                    DOMSelectElement select = (DOMSelectElement) document.findElement(By.id("select-tag"));
-                    selectOptionByIndex(select, 2);
+                    DOMDocument document = event.getBrowser().getDocument();
+                    DOMElement link = document.findElement(By.id("link"));
+                    Map<String, String> attributes = link.getAttributes();
+                    for (String attrName : attributes.keySet()) {
+                        System.out.println(attrName + " = " + attributes.get(attrName));
+                    }
                 }
             }
         });
-        browser.loadHTML("<html><body><select id='select-tag'>\n" +
-                "  <option value=\"volvo\">Volvo</option>\n" +
-                "  <option value=\"saab\">Saab</option>\n" +
-                "  <option value=\"opel\">Opel</option>\n" +
-                "  <option value=\"audi\">Audi</option>\n" +
-                "</select></body></html>");
-    }
-
-    private static void selectOptionByIndex(DOMSelectElement select, int index) {
-        List<DOMOptionElement> options = select.getOptions();
-        options.get(2).setSelected(true);
+        browser.loadHTML("<html><body><a href='#' id='link' title='link title'></a></body></html>");
     }
 }
