+++ dom/elements/src/main/java/DOMQuerySelector.java
@@ -22,25 +22,25 @@ import com.teamdev.jxbrowser.chromium.Browser;
 import com.teamdev.jxbrowser.chromium.dom.By;
 import com.teamdev.jxbrowser.chromium.dom.DOMDocument;
 import com.teamdev.jxbrowser.chromium.dom.DOMElement;
+import com.teamdev.jxbrowser.chromium.dom.DOMNode;
 import com.teamdev.jxbrowser.chromium.events.FinishLoadingEvent;
 import com.teamdev.jxbrowser.chromium.events.LoadAdapter;
 import com.teamdev.jxbrowser.chromium.swing.BrowserView;
 
 import javax.swing.*;
 import java.awt.*;
-import java.util.Map;
+import java.util.List;
 
 /**
- * Demonstrates how to get list of existing attributes of a specified HTML element.
+ * This example demonstrates how to use querySelector DOM API.
  */
-public class DOMGetAttributesExample {
+public class DOMQuerySelector {
     public static void main(String[] args) {
         Browser browser = new Browser();
         BrowserView browserView = new BrowserView(browser);
 
         JFrame frame = new JFrame();
-        frame.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
-        frame.add(browserView, BorderLayout.CENTER);
+        frame.getContentPane().add(browserView, BorderLayout.CENTER);
         frame.setSize(800, 600);
         frame.setLocationRelativeTo(null);
         frame.setVisible(true);
@@ -48,16 +48,22 @@ public class DOMGetAttributesExample {
         browser.addLoadListener(new LoadAdapter() {
             @Override
             public void onFinishLoadingFrame(FinishLoadingEvent event) {
-                if (event.isMainFrame()) {
-                    DOMDocument document = event.getBrowser().getDocument();
-                    DOMElement link = document.findElement(By.id("link"));
-                    Map<String, String> attributes = link.getAttributes();
-                    for (String attrName : attributes.keySet()) {
-                        System.out.println(attrName + " = " + attributes.get(attrName));
-                    }
+                DOMDocument document = event.getBrowser().getDocument();
+                // Get the div with id = "root".
+                DOMNode divRoot = document.findElement(By.cssSelector("#root"));
+                // Get all paragraphs.
+                List<DOMElement> paragraphs = divRoot.findElements(By.cssSelector("p"));
+                for (DOMElement paragraph : paragraphs) {
+                    System.out.println("paragraph.getNodeValue() = " +
+                            paragraph.getNodeValue());
                 }
             }
         });
-        browser.loadHTML("<html><body><a href='#' id='link' title='link title'></a></body></html>");
+        browser.loadHTML(
+                "<html><body><div id='root'>" +
+                        "<p>paragraph1</p>" +
+                        "<p>paragraph2</p>" +
+                        "<p>paragraph3</p>" +
+                        "</div></body></html>");
     }
 }
